Add files via upload
This is a part of the phase 1 source code that I have added to the project. A study option should also be added to the JMenu. That shall be done in the future. The next part shall entail adding more functionality to each of the JMenuItems when they are selected/clicked.  Implementations for two classes have also been provided, one for the Deck class and another for the FlashCard class.